creacion del combine reducer, cambio en archivos
