+++ client/src/Redux/actions-types/index.js
@@ -1,12 +1,10 @@
-import axios from 'axios';
-import { SIGN_UP } from '../actions/index';
+// En este archivo se hacen las importaciones de todos los actions types
+// Definir el nomnre de los archivos como: nombre/funcion + Actions
 
-export function signUpUser(data) {
-  return function(dispatch) {
-    return axios.post(`http://localhost:3001/users`, data)
-      .then(response => {
-        dispatch({ type: SIGN_UP, payload: response.data });
-      })
-      .catch(error => console.log(error));
-  };
+import signUpUserActions from "./signUpUserActions";
+
+const allActions = {
+  signUpUserActions,
 };
+
+export default allActions;
