+++ client/src/Redux/actions/index.js
@@ -1,1 +1,3 @@
+// definir todas las acciones en este archivo
+
 export const SIGN_UP = 'SIGN_UP';
